+++ src/components/AppHeader/AppHeader.module.css
@@ -38,7 +38,7 @@
     /* navigation */
     
     position: relative;
-    width: 80%;
+    width: 1240px;
     height: 56px;
 
     padding: 0px 326px;
